# server/openapi-with-docs.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/connect": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Connect",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Register",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Login",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/join": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Join",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/send": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Send",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/poke": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Poke",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/leave": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Leave",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/typing": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Typing",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypingRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/rooms": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_CreateRoom",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Messages": {
      "post": {
        "tags": [
          "Messages"
        ],
        "operationId": "Messages_SendMessage",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Messages/{roomName}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "Messages_GetLastMessages",
        "parameters": [
          {
            "name": "roomName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "JoinDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "room": {
            "type": "string"
          }
        }
      },
      "ChatMessageDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "room": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "TypingRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "room": {
            "type": "string"
          },
          "isTyping": {
            "type": "boolean"
          }
        }
      },
      "CreateRoomDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "roomName": {
            "type": "string"
          }
        }
      },
      "CreateMessageDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string"
          },
          "roomName": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "recipientUserId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      }
    }
  }
}